:truck: move: 응답 DTO로 매핑을 컨트롤러 레이어가 하도록 변경한다
