+++ api/src/main/java/com/pancake/api/watch/api/WatchContentResponse.java
@@ -1,4 +1,4 @@
-package com.pancake.api.watch.application;
+package com.pancake.api.watch.api;
 
 import com.pancake.api.watch.domain.WatchContent;
 import com.pancake.api.watch.domain.WatchOption;
@@ -9,7 +9,7 @@ import java.util.List;
 
 @Data
 @NoArgsConstructor
-public class GetContentsToWatchResult {
+public class WatchContentResponse {
 
     private Long id;
     private String title;
@@ -17,9 +17,9 @@ public class GetContentsToWatchResult {
     private String imageUrl;
     private boolean watched;
 
-    private List<OptionResult> options;
+    private List<Option> options;
 
-    public GetContentsToWatchResult(WatchContent watchContent) {
+    public WatchContentResponse(WatchContent watchContent) {
         this.id = watchContent.getId();
         this.title = watchContent.getTitle();
         this.description = watchContent.getDescription();
@@ -28,17 +28,17 @@ public class GetContentsToWatchResult {
         this.options = toOptions(watchContent.getOptions());
     }
 
-    private static List<OptionResult> toOptions(List<WatchOption> watchingOptions) {
-        return watchingOptions.stream().map(OptionResult::new).toList();
+    private static List<Option> toOptions(List<WatchOption> watchingOptions) {
+        return watchingOptions.stream().map(Option::new).toList();
     }
 
     @Data
     @NoArgsConstructor
-    public static class OptionResult {
+    public static class Option {
         private String platformName;
         private String platformLabel;
 
-        public OptionResult(WatchOption option) {
+        public Option(WatchOption option) {
             this.platformName = option.getPlatform().name();
             this.platformLabel = option.getPlatform().label();
         }
