+++ api/src/test/java/com/pancake/api/watch/api/WatchApiControllerTest.java
@@ -1,8 +1,6 @@
 package com.pancake.api.watch.api;
 
 import com.pancake.api.watch.application.GetContentsToWatch;
-import com.pancake.api.watch.application.GetContentsToWatchResult;
-import lombok.Builder;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
@@ -11,6 +9,7 @@ import org.springframework.test.web.reactive.server.WebTestClient;
 
 import java.util.List;
 
+import static com.pancake.api.watch.Builders.aWatchContent;
 import static org.mockito.BDDMockito.given;
 
 @WebMvcTest(WatchApiController.class)
@@ -25,7 +24,7 @@ class WatchApiControllerTest {
     @Test
     void getAll() {
         //given
-        var expected = List.of(aResult().id(1L).build(), aResult().id(2L).build());
+        var expected = List.of(aWatchContent().id(1L).build(), aWatchContent().id(2L).build());
 
         given(getContentsToWatch.query()).willReturn(expected);
 
@@ -35,15 +34,7 @@ class WatchApiControllerTest {
         //then
         response.expectAll(
                 spec -> spec.expectStatus().isOk(),
-                spec -> spec.expectBodyList(GetContentsToWatchResult.class).isEqualTo(expected)
+                spec -> spec.expectBodyList(WatchContentResponse.class).hasSize(2)
         );
-    }
-
-    @Builder(builderMethodName = "aResult")
-    private static GetContentsToWatchResult create(long id) {
-        var result = new GetContentsToWatchResult();
-        result.setId(id);
-
-        return result;
     }
 }
